Ya esta todo el backend hecho y subido espero este bien

+++ src/main/java/com/gesinv/gestioninventario/dto/CategoriaDto.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Setter
 @Getter
 public class CategoriaDto {
-    private Long idCategoria;
+    private Integer idCategoria;
     private String descripcion;
 }
