+++ src/main/java/com/gesinv/gestioninventario/dto/MarcaDto.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Setter
 @Getter
 public class MarcaDto {
-    private Long idMarca;
+    private Integer idMarca;
     private String descripcion;
 }
