+++ src/main/java/com/gesinv/gestioninventario/dto/RolDto.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 
 public class RolDto {
-    private Long idRol;
+    private Integer idRol;
     private String nombre;
 }
