+++ src/main/java/com/gesinv/gestioninventario/repository/ICategoriaRepo.java
@@ -3,7 +3,7 @@ package com.gesinv.gestioninventario.repository;
 import com.gesinv.gestioninventario.modelo.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICategoriaRepo extends JpaRepository<Categoria, Long> {
+public interface ICategoriaRepo extends JpaRepository<Categoria, Integer> {
 
 }
