+++ src/main/java/com/gesinv/gestioninventario/repository/IMarcaRepo.java
@@ -3,7 +3,7 @@ package com.gesinv.gestioninventario.repository;
 import com.gesinv.gestioninventario.modelo.Marca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IMarcaRepo extends JpaRepository<Marca, Long> {
+public interface IMarcaRepo extends JpaRepository<Marca, Integer> {
 
 
 }
