+++ src/main/java/com/gesinv/gestioninventario/repository/IProductoRepo.java
@@ -3,6 +3,6 @@ package com.gesinv.gestioninventario.repository;
 import com.gesinv.gestioninventario.modelo.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IProductoRepo extends JpaRepository<Producto, Long> {
+public interface IProductoRepo extends JpaRepository<Producto, Integer> {
 
 }
