+++ src/main/java/com/gesinv/gestioninventario/repository/IRolRepo.java
@@ -3,7 +3,7 @@ package com.gesinv.gestioninventario.repository;
 import com.gesinv.gestioninventario.modelo.Rol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRolRepo extends JpaRepository<Rol, Long> {
+public interface IRolRepo extends JpaRepository<Rol, Integer> {
 
 }
