+++ src/main/java/com/gesinv/gestioninventario/repository/IUsuarioRepo.java
@@ -3,7 +3,7 @@ package com.gesinv.gestioninventario.repository;
 import com.gesinv.gestioninventario.modelo.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUsuarioRepo extends JpaRepository<Usuario, Long> {
+public interface IUsuarioRepo extends JpaRepository<Usuario, Integer> {
 
 
 }
