+++ src/main/java/com/gesinv/gestioninventario/repository/IZonaRepo.java
@@ -3,6 +3,6 @@ package com.gesinv.gestioninventario.repository;
 import com.gesinv.gestioninventario.modelo.Zona;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IZonaRepo extends JpaRepository<Zona, Long> {
+public interface IZonaRepo extends JpaRepository<Zona, Integer> {
 
 }
